refactor: improve separation of core packages
