+++ elasticai/creator/file_generation/in_memory_path.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
-from elasticai.creator.vhdl.code_generation.template import Template, TemplateExpander
-from elasticai.creator.vhdl.savable import File, Path
+from elasticai.creator.file_generation.savable import File, Path
+from elasticai.creator.templating.template import Template, TemplateExpander
 
 
 class InMemoryFile(File):
