+++ elasticai/creator/file_generation/resource_utils.py
