+++ elasticai/creator/nn/hard_sigmoid/design.py
@@ -1,10 +1,7 @@
+from elasticai.creator.file_generation.savable import Path
+from elasticai.creator.templating.template import InProjectTemplate, module_to_package
 from elasticai.creator.vhdl.auto_wire_protocols.port_definitions import create_port
-from elasticai.creator.vhdl.code_generation.template import (
-    InProjectTemplate,
-    module_to_package,
-)
 from elasticai.creator.vhdl.design.design import Design, Port
-from elasticai.creator.vhdl.savable import Path
 
 
 class FPHardSigmoid(Design):
