+++ elasticai/creator/nn/identity/design.py
@@ -1,12 +1,9 @@
+from elasticai.creator.file_generation.savable import Path
+from elasticai.creator.templating.template import InProjectTemplate, module_to_package
 from elasticai.creator.vhdl.auto_wire_protocols.port_definitions import create_port
 from elasticai.creator.vhdl.code_generation.addressable import calculate_address_width
-from elasticai.creator.vhdl.code_generation.template import (
-    InProjectTemplate,
-    module_to_package,
-)
 from elasticai.creator.vhdl.design.design import Design
 from elasticai.creator.vhdl.design.ports import Port
-from elasticai.creator.vhdl.savable import Path
 
 
 class BufferedIdentity(Design):
