+++ elasticai/creator/nn/linear/design.py
@@ -1,14 +1,11 @@
 from itertools import chain
 from typing import Optional
 
+from elasticai.creator.file_generation.savable import Path
+from elasticai.creator.templating.template import InProjectTemplate, module_to_package
 from elasticai.creator.vhdl.auto_wire_protocols.port_definitions import create_port
-from elasticai.creator.vhdl.code_generation.template import (
-    InProjectTemplate,
-    module_to_package,
-)
 from elasticai.creator.vhdl.design.design import Design
 from elasticai.creator.vhdl.design.ports import Port
-from elasticai.creator.vhdl.savable import Path
 from elasticai.creator.vhdl.shared_designs.rom import Rom
