+++ elasticai/creator/nn/sequential/design.py
@@ -1,6 +1,8 @@
 from functools import partial, reduce
 from itertools import chain
 
+from elasticai.creator.file_generation.savable import Path
+from elasticai.creator.templating.template import InProjectTemplate, module_to_package
 from elasticai.creator.vhdl.auto_wire_protocols.autowiring import (
     AutoWirer,
     DataFlowNode,
@@ -11,13 +13,8 @@ from elasticai.creator.vhdl.code_generation.code_abstractions import (
     create_instance,
     create_signal_definitions,
 )
-from elasticai.creator.vhdl.code_generation.template import (
-    InProjectTemplate,
-    module_to_package,
-)
 from elasticai.creator.vhdl.design.design import Design
 from elasticai.creator.vhdl.design.ports import Port
-from elasticai.creator.vhdl.savable import Path
 
 
 class Sequential(Design):
