+++ elasticai/creator/vhdl/shared_designs/rom/design.py
@@ -1,14 +1,11 @@
 from functools import partial
 
+from elasticai.creator.file_generation.savable import Path
+from elasticai.creator.templating.template import InProjectTemplate, module_to_package
 from elasticai.creator.vhdl.code_generation.addressable import calculate_address_width
 from elasticai.creator.vhdl.code_generation.code_abstractions import (
     to_vhdl_binary_string,
 )
-from elasticai.creator.vhdl.code_generation.template import (
-    InProjectTemplate,
-    module_to_package,
-)
-from elasticai.creator.vhdl.savable import Path
 
 
 class Rom:
