+++ tests/nn/sequential/test_sequential_signal_widths_are_derived_from_subdesigns.py
@@ -1,10 +1,10 @@
 import pytest
 
+from elasticai.creator.file_generation.savable import Path
 from elasticai.creator.nn.sequential.design import Sequential
 from elasticai.creator.vhdl.auto_wire_protocols.port_definitions import create_port
 from elasticai.creator.vhdl.design.design import Design
 from elasticai.creator.vhdl.design.ports import Port
-from elasticai.creator.vhdl.savable import Path
 
 
 class DummyDesign(Design):
