+++ tests/vhdl/designs/rom/test_rom_file.py
@@ -4,7 +4,7 @@ from typing import cast
 
 import pytest
 
-from elasticai.creator.in_memory_path import InMemoryFile, InMemoryPath
+from elasticai.creator.file_generation.in_memory_path import InMemoryFile, InMemoryPath
 from elasticai.creator.vhdl.shared_designs.rom import Rom
 
 from .utils import extract_rom_values
